requirements.txt
django
mysqlclient
pillow
django-ninja
django-cors-headers

django-redis
django-celery-results
django-celery-beat
django-inlinecss

celery==5.2.3
redis